feat: complete basic game loop
